requirements.txt
customtkinter
pillow
vaderSentiment
matplotlib
wordcloud
reportlab